requirements_dev.txt
plyfile
tqdm
einops
lpips
loguru
simple-parsing
pycolmap
ninja
matplotlib
colorama
vector-quantize-pytorch==1.22.0
dahuffman
torchac
constriction
scikit-learn
opencv-python
point_cloud_utils
trimesh
open3d
kornia
tensorboard
numpy==1.26.4
